fix: cast the staleness bound, or the health endpoint 500s
43100863 shipped a live 500 on /api/reconciliation:

    ERROR: operator does not exist: text * interval (SQLSTATE 42883)

The split staleness bound is
`CASE WHEN status = $3 THEN $2 ELSE $1 END * INTERVAL '1 second'`. The driver
sends those parameters untyped, so Postgres infers the CASE result as `text`,
and `text * interval` is not an operator. Every row the query would have
returned became an error instead — the stuck-order half of the health board,
including the queued anomaly 43100863 was written to add.

Fixed by casting both branches: `$2::int` / `$1::int`.

WHY IT GOT THROUGH, since the query WAS checked against the live plant database
before deploying. The check was `PREPARE stuckq(int, int, text)` in psql, which
passes — because declaring the parameter types is exactly what the driver does
not do. A validation that supplies the missing information cannot detect that it
is missing.

So the test added here goes through the DRIVER, against real Postgres, which is
the only check that reproduces the failure. It is nominally a threshold test and
is really an "does this query execute at all" test; the threshold assertions are
the excuse to run it. Verified in both directions — with the casts removed it
fails on the exact SQLSTATE above.

No existing test covered ListAnomalies against Postgres at all: the engine tests
delegate to a fake, and the store package's docker tests cover
ListOrderCompletionAnomalies, a different query. That gap is what made a
runtime-only SQL error invisible to build, vet, and the full -race suite.

+++ shingo-core/store/reconciliation/reconciliation.go
@@ -208,12 +208,21 @@ func ListAnomalies(db *sql.DB) ([]*Anomaly, error) {
 	// Splitting it in SQL rather than running two queries keeps the ORDER BY over
 	// the whole result, so the oldest anomaly is still first regardless of which
 	// bound produced it.
+	//
+	// THE ::int CASTS ARE LOAD-BEARING. The driver sends these parameters
+	// untyped, so without them Postgres infers the CASE result as `text` and the
+	// whole query dies at RUNTIME on `operator does not exist: text * interval`
+	// — a live 500 on the health endpoint, from a query that builds and vets
+	// clean. It also survives a psql `PREPARE stuckq(int, int, text)` check,
+	// because declaring the types is exactly what the driver does not do.
+	// TestListAnomalies_QueuedGetsTheLongerBound exercises this through the
+	// driver, which is the only check that would have caught it.
 	rows, err := db.Query(fmt.Sprintf(`
 		SELECT id, status, updated_at
 		FROM orders
 		WHERE status IN (%s)
 		  AND updated_at < NOW() - (
-		        CASE WHEN status = $3 THEN $2 ELSE $1 END * INTERVAL '1 second')
+		        CASE WHEN status = $3 THEN $2::int ELSE $1::int END * INTERVAL '1 second')
 		ORDER BY updated_at ASC`, protocol.RuntimeStuckCandidateStatusSQLList()),
 		int(stuckOrderAge.Seconds()), int(queuedOrderAge.Seconds()), string(protocol.StatusQueued))
 	if err != nil {
